Use mobx store for dapps
