+++ js/src/views/Dapps/dapps.js
@@ -15,45 +15,39 @@
 // along with Parity.  If not, see <http://www.gnu.org/licenses/>.
 
 import React, { Component, PropTypes } from 'react';
+import { observer } from 'mobx-react';
 
-import Contracts from '../../contracts';
-import { hashToImageUrl } from '../../redux/util';
 import { Actionbar, Page } from '../../ui';
 import FlatButton from 'material-ui/FlatButton';
 import EyeIcon from 'material-ui/svg-icons/image/remove-red-eye';
 
-import { fetchAvailable } from './registry';
-import { readHiddenApps, writeHiddenApps } from './hidden';
+import DappsStore from './dappsStore';
 
 import AddDapps from './AddDapps';
 import Summary from './Summary';
 
 import styles from './dapps.css';
 
+@observer
 export default class Dapps extends Component {
   static contextTypes = {
     api: PropTypes.object.isRequired
   }
 
+  store = new DappsStore(this.context.api);
+
   state = {
-    available: [],
-    hidden: [],
     modalOpen: false
   }
 
-  componentDidMount () {
-    this.loadAvailableApps();
-  }
-
   render () {
-    const { available, hidden, modalOpen } = this.state;
-    const apps = available.filter((app) => !hidden.includes(app.id));
+    const { modalOpen } = this.state;
 
     return (
       <div>
         <AddDapps
-          available={ available }
-          hidden={ hidden }
+          available={ this.store.apps }
+          hidden={ this.store.hidden }
           open={ modalOpen }
           onHideApp={ this.onHideApp }
           onShowApp={ this.onShowApp }
@@ -68,7 +62,7 @@ export default class Dapps extends Component {
         />
         <Page>
           <div className={ styles.list }>
-            { apps.map(this.renderApp) }
+            { this.store.visibleApps.map(this.renderApp) }
           </div>
         </Page>
       </div>
@@ -90,19 +84,11 @@ export default class Dapps extends Component {
   }
 
   onHideApp = (id) => {
-    const { hidden } = this.state;
-    const newHidden = hidden.concat(id);
-
-    this.setState({ hidden: newHidden });
-    writeHiddenApps(newHidden);
+    this.store.hideApp(id);
   }
 
   onShowApp = (id) => {
-    const { hidden } = this.state;
-    const newHidden = hidden.filter((_id) => _id !== id);
-
-    this.setState({ hidden: newHidden });
-    writeHiddenApps(newHidden);
+    this.store.showApp(id);
   }
 
   openModal = () => {
@@ -112,57 +98,4 @@ export default class Dapps extends Component {
   closeModal = () => {
     this.setState({ modalOpen: false });
   };
-
-  loadAvailableApps () {
-    const { api } = this.context;
-
-    fetchAvailable(api)
-      .then((available) => {
-        this.setState({
-          available,
-          hidden: readHiddenApps()
-        });
-
-        this.loadContent();
-      });
-  }
-
-  loadContent () {
-    const { api } = this.context;
-    const { available } = this.state;
-    const { dappReg } = Contracts.get();
-
-    return Promise
-      .all(available.map((app) => dappReg.getImage(app.id)))
-      .then((images) => {
-        const _available = images
-          .map(hashToImageUrl)
-          .map((image, index) => Object.assign({}, available[index], { image }));
-
-        this.setState({ available: _available });
-        const _networkApps = _available.filter((app) => app.network);
-
-        return Promise
-          .all(_networkApps.map((app) => dappReg.getContent(app.id)))
-          .then((content) => {
-            const networkApps = content.map((_contentHash, index) => {
-              const networkApp = _networkApps[index];
-              const contentHash = api.util.bytesToHex(_contentHash).substr(2);
-              const app = _available.find((_app) => _app.id === networkApp.id);
-
-              console.log(`found content for ${app.id} at ${contentHash}`);
-              return Object.assign({}, app, { contentHash });
-            });
-
-            this.setState({
-              available: _available.map((app) => {
-                return Object.assign({}, networkApps.find((napp) => app.id === napp.id) || app);
-              })
-            });
-          });
-      })
-      .catch((error) => {
-        console.warn('loadImages', error);
-      });
-  }
 }
